fix(context): carry the pipeline score onto every bundled memory row (#155)
## What

Every memory row in a `comemory context` bundle now carries the
retrieval pipeline's `final_score` for that memory, on both the CLI and
`GET|POST /api/v1/context`. Before, every row reported `score: 0.0` (the
documented "caller-supplied" default that no caller supplied), so a
console or agent reading the bundle could not tell the top-ranked memory
from the last one.

Closes #152.

## How

- `retrieval::bundle::assemble` takes `&[RankedMemory]` (id + score)
instead of `&[String]`. `RankedMemory: From<Reranked>` carries
`parts.final_score` across, and `collect_memory` writes it onto the row.
- `api::context::run` maps `run.hits` through it; nothing else in the
pipeline changes, the ordering was already right.
- `MemoryBundleRow::score` is now `f64`, the pipeline's native type, so
the bundle reports exactly the number `search --json` reports as each
hit's `score` for the same query. `tty::score` takes `f64` too, which
drops the `as f32` cast the `search` / `search-code` renderers were
doing.

## Tests (real data, no mocks)

-
`src/retrieval/tests/bundle.rs::memory_rows_carry_the_ranked_score_in_caller_order`
— the scores land on the rows in the caller's order and reach the JSON
verbatim.
- `tests/cli__context.rs::context_memory_rows_carry_the_pipeline_score`
— two saved memories; every bundled row's score is `> 0`,
non-increasing, and equals `comemory search --json`'s per-hit `score`
for the same query (tracking off on both runs).
-
`tests/api__context.rs::run_carries_the_search_pipeline_score_onto_each_memory_row`
— the same pin over `api::search::run` / `api::context::run` directly.

## Docs

`docs/scenarios/context.md` (context-01 expectation),
`src/retrieval/README.md`.

`bash scripts/check-all.sh` green; `cargo nextest run --all-features`
2314/2314.

+++ src/output/search_code.rs
@@ -140,7 +140,7 @@ pub fn write_tty(
         writeln!(
             out,
             "{}  {}:{}-{}  {}  ({})  {}",
-            tty::score(hit.parts.final_score as f32),
+            tty::score(hit.parts.final_score),
             hit.path,
             hit.line_start,
             hit.line_end,
